Add pagination to tm bot ui
