+++ pkg/tm-bot/ui/pages/command_help.go
