+++ pkg/tm-bot/ui/static/main.css
@@ -3,3 +3,4 @@
 @import "commandhelp.css";
 @import "pr-status.css";
 @import "testrun.css";
+@import "pagination.css";
